Update .gitignore
Reorganize structure so it's by alphabetical order.
Added the important files that we do not want to have tracked by GIT when we first clone the repository. This is to prevent problems in the future for when a user fetches the repositories modifications.
When using git command-line under linux, normal behavior is to use: git pull from time to time to update source files to the latest.
If we have thoses files tracked by git, it will OVERWRITE them making faveo UNUSABLE for the customer/client.
This is something you have not put before because most of you are running under windows/mac i assume and are not in a position or being a simple user. Therefore, when you are testing you test as developpers not as users. I noticed that bug before and did a pull request but it seems nobody took care of that or looked at it seriously.