Bug-fix-patch-16
Starting count for bug patch so numbered this pacth as 16 (I hope we've
commit atleast 15 bug fixes till now)
Solved bugs:
*While replying on a ticket as an agent subject of a mail is getting
changed which results in new mail thread creation in emails.
*Ticket title sometimes get replaced by ''(an empty string) because
sometimes it fetches last created thread as first row and it my not have
the subject/title. So tried to fetch only first row which contains
subjetc of the tickets.