Fixing new ticket creation bug upon user replay in emails with Farsi or Arabic subject
Everything works perfect till the user send an email and you answer it. when the user replays to your answer, if the subject is in Farsi or Arabic it will create a new ticket. this is not the intended behavior. the bug was in `TicketController` where the `check_ticket` function didn't take MIME headers for Farsi and Arabic into account. now the bug has been resolved and Faveo-HelpDesk is working perfectly.

also we fixed the styling